requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-admin==2.0.2
django-cors-headers==4.4.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
drf-yasg==1.21.7
excel-base==1.0.4
idna==3.7
inflection==0.5.1
isoweek==1.3.3
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
screen==1.0.1
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
xlwt==1.3.0
